[#10] Create simulation time system [Part 1 - times of day]
Related issue: https://github.com/HikariIT/Foxes-formal-model/issues/10
Description:

Added TimeManager to the project. It is responsible for keeping track of the current time of day
and the current day of the simulation. It calculates the time of day based on the current time of simulation,
location and the time of sunrise and sunset.